Move debug generator and store entries in separate JSON files
